+++ jet/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'jet.apps.JetConfig'
 
-__version__ = '1.0.9-alpha.6'
+__version__ = '1.0.9-alpha.7'
 VERSION = __version__